Re-introduce SHA-256
This may be required for a future refactoring to openssl and away from
libsodium. The SHA-256 sums are only generated, but not verified in
this commit.

+++ src/sparse_file.c
@@ -363,7 +363,7 @@ int sparse_file_is_supported(const char *testfile, size_t min_hole_size)
         ERR_CLEANUP_MSG("Sparse check write to offset %d failed.", min_hole_size);
 
     lseek(fd, 0, SEEK_SET);
-    off_t offset = lseek(fd, 0, SEEK_DATA);
+    size_t offset = lseek(fd, 0, SEEK_DATA);
     if (offset != min_hole_size)
         ERR_CLEANUP_MSG("Hole of %d bytes not created on filesystem", min_hole_size);
